delivery edit view, apply stored referer url changes

+++ lib/pack_material/interactors/mr_delivery_interactor.rb
@@ -15,29 +15,39 @@ module PackMaterialApp
     end
 
     def create_mr_delivery(params)
-      res = validate_mr_delivery_params(params)
-      return validation_failed_response(res) unless res.messages.empty?
-      id = nil
-      repo.transaction do
-        id = repo.create_mr_delivery(res)
-        log_status('mr_deliveries', id, 'CREATED')
-        log_transaction
+      can_create = TaskPermissionCheck::MrDelivery.call(:create)
+      if can_create.success
+        res = validate_mr_delivery_params(params)
+        return validation_failed_response(res) unless res.messages.empty?
+        id = nil
+        repo.transaction do
+          id = repo.create_mr_delivery(res)
+          log_status('mr_deliveries', id, 'CREATED')
+          log_transaction
+        end
+        instance = mr_delivery(id)
+        success_response("Created delivery #{instance.delivery_number}", instance)
+      else
+        failed_response(can_create.message)
       end
-      instance = mr_delivery(id)
-      success_response("Created delivery #{instance.delivery_number}", instance)
     rescue Sequel::UniqueConstraintViolation
       validation_failed_response(OpenStruct.new(messages: { base: ['This delivery already exists'] }))
     end
 
     def update_mr_delivery(id, params)
-      res = validate_mr_delivery_params(params)
-      return validation_failed_response(res) unless res.messages.empty?
-      repo.transaction do
-        repo.update_mr_delivery(id, res)
-        log_transaction
+      can_update = TaskPermissionCheck::MrDelivery.call(:update, id)
+      if can_update.success
+        res = validate_mr_delivery_params(params)
+        return validation_failed_response(res) unless res.messages.empty?
+        repo.transaction do
+          repo.update_mr_delivery(id, res)
+          log_transaction
+        end
+        instance = mr_delivery(id)
+        success_response("Updated delivery #{instance.delivery_number}", instance)
+      else
+        failed_response(can_update.message)
       end
-      instance = mr_delivery(id)
-      success_response("Updated delivery #{instance.delivery_number}", instance)
     end
 
     def verify_mr_delivery(id) # rubocop:disable Metrics/AbcSize
@@ -57,13 +67,18 @@ module PackMaterialApp
     end
 
     def delete_mr_delivery(id)
-      name = mr_delivery(id).delivery_number
-      repo.transaction do
-        repo.delete_mr_delivery(id)
-        log_status('mr_deliveries', id, 'DELETED')
-        log_transaction
+      can_delete = TaskPermissionCheck::MrDelivery.call(:delete, id)
+      if can_delete.success
+        name = mr_delivery(id).delivery_number
+        repo.transaction do
+          repo.delete_mr_delivery(id)
+          log_transaction
+          log_status('mr_deliveries', id, 'DELETED')
+        end
+        success_response("Deleted delivery #{name}")
+      else
+        failed_response(can_delete.message)
       end
-      success_response("Deleted delivery #{name}")
     end
   end
 end
