+++ lib/pack_material/interactors/mr_delivery_item_interactor.rb
@@ -16,41 +16,54 @@ module PackMaterialApp
 
     def create_mr_delivery_item(parent_id, params)
       params[:mr_delivery_id] = parent_id
-      res = validate_mr_delivery_item_params(params)
-      return validation_failed_response(res) unless res.messages.empty?
-      id = nil
-      repo.transaction do
-        id = repo.create_mr_delivery_item(res)
-        log_status('mr_delivery_items', id, 'CREATED')
-        log_transaction
+      can_create = TaskPermissionCheck::MrDeliveryItem.call(:create, delivery_id: parent_id)
+      if can_create.success
+        res = validate_mr_delivery_item_params(params)
+        return validation_failed_response(res) unless res.messages.empty?
+        id = nil
+        repo.transaction do
+          id = repo.create_mr_delivery_item(res)
+          log_status('mr_delivery_items', id, 'CREATED')
+          log_transaction
+        end
+        instance = mr_delivery_item(id)
+        success_response("Created delivery item #{instance.remarks}", instance)
+      else
+        failed_response(can_create.message)
       end
-      instance = mr_delivery_item(id)
-      success_response("Created delivery item #{instance.remarks}",
-                       instance)
     rescue Sequel::UniqueConstraintViolation
       validation_failed_response(OpenStruct.new(messages: { remarks: ['This delivery item already exists'] }))
     end
 
     def update_mr_delivery_item(id, params)
-      res = validate_mr_delivery_item_params(params)
-      return validation_failed_response(res) unless res.messages.empty?
-      repo.transaction do
-        repo.update_mr_delivery_item(id, res)
-        log_transaction
+      can_update = TaskPermissionCheck::MrDeliveryItem.call(:update, id)
+      if can_update.success
+        res = validate_mr_delivery_item_params(params)
+        return validation_failed_response(res) unless res.messages.empty?
+        repo.transaction do
+          repo.update_mr_delivery_item(id, res)
+          log_transaction
+        end
+        instance = mr_delivery_item(id)
+        success_response("Updated delivery item #{instance.remarks}", instance)
+      else
+        failed_response(can_update.message)
       end
-      instance = mr_delivery_item(id)
-      success_response("Updated delivery item #{instance.remarks}",
-                       instance)
     end
 
     def delete_mr_delivery_item(id)
-      name = mr_delivery_item(id).remarks
-      repo.transaction do
-        repo.delete_mr_delivery_item(id)
-        log_status('mr_delivery_items', id, 'DELETED')
-        log_transaction
+      can_delete = TaskPermissionCheck::MrDeliveryItem.call(:delete, id)
+      if can_delete.success
+        name = mr_delivery_item(id).remarks
+        repo.transaction do
+          repo.delete_mr_delivery_item(id)
+          log_status('mr_delivery_items', id, 'DELETED')
+          log_transaction
+        end
+        success_response("Deleted delivery item #{name}")
+      else
+        failed_response(can_delete.message)
       end
-      success_response("Deleted delivery item #{name}")
     end
 
     def available_purchase_order_items(purchase_order_id, delivery_id)
